requirements.txt
environs==14.3.0
marshmallow==4.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
